Migrate Mutation.schemaDelete
